ai-skincare-backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
pymongo==4.3.3
Werkzeug==2.2.3
python-dotenv==1.0.0
gunicorn==20.1.0
Pillow==9.4.0
email-validator==1.3.1
pytest==7.3.1 